+++ backend/src/login.rs
@@ -1,15 +1,16 @@
-use chrono::Utc;
-use actix_web::error::ErrorInternalServerError;
-use actix_web::{HttpResponse, Responder, post, web, get};
-use base64::{self, Engine as _};
-use crypto::common::typenum::True;
+use std::result;
+
+use actix_web::{get, post, web, HttpMessage, HttpResponse, Responder};
+use futures_util::future::UnwrapOrElse;
 use serde::{Serialize, Deserialize};
 use rand::RngCore;
-use actix_web::web::Data;
-use actix_web::web::Json;
 use sqlx::PgPool;
 use sqlx::Error;
-use sqlx::Row;
+use actix_web::HttpRequest;
+use actix_identity::Identity;
+use base64::{engine::general_purpose::URL_SAFE, Engine};
+
+
 
 #[derive(Serialize, Deserialize)]
 pub struct LoginData {
@@ -19,7 +20,7 @@ pub struct LoginData {
 }
 
 #[post("/login")]
-pub async fn login(pool: web::Data<PgPool>, data: web::Json<LoginData>) -> impl Responder {
+pub async fn login(pool: web::Data<PgPool>, data: web::Json<LoginData>, request: HttpRequest,) -> impl Responder {
     // Query the database to check if the user exists
     let user_result = sqlx::query(
         "SELECT * FROM users WHERE username = $1 AND password = $2",
@@ -31,26 +32,18 @@ pub async fn login(pool: web::Data<PgPool>, data: web::Json<LoginData>) -> impl 
     .await;
 
     match user_result {
-        Ok(user) => {
-            let token = gentoken().await;
-            let time_created = Utc::now();
-            println!("Token: {:?}", token);
-            _ = sqlx::query(
-                "INSERT INTO tokens (username, token, timecreated) VALUES ($1, $2, $3)",
-            )
-            .bind(&data.username)
-            .bind(&token)
-            .bind(time_created.to_rfc3339())
-            .execute(pool.get_ref())
-            .await
-            .map_err(|e| ErrorInternalServerError(e));
-
-            HttpResponse::Ok().json(token)
+        Ok(_user_result) => {
+           let mut random_bytes = [0u8; 32];
+            rand::rng().fill_bytes(&mut random_bytes);
+            let session_token = URL_SAFE.encode(&random_bytes);
+            Identity::login(&request.extensions(),session_token.clone()).unwrap();
+            println!("Provisioned session token: {}", session_token);
+            HttpResponse::Ok().body("Session token provisioned")
         }
         Err(Error::RowNotFound) => {
             println!("Invalid username or password");
             HttpResponse::Unauthorized().body("Invalid username or password")
-       }
+        }
         Err(e) => {
             println!("Internal server error: {}", e);
             HttpResponse::InternalServerError().body(format!("Internal server error: {}", e))
@@ -98,23 +91,6 @@ pub async fn checktoken(pool: web::Data<PgPool>, data: web::Json<TokenData>) -> 
         }
     }
 }*/
-async fn gentoken() -> String {
-// TODO: REMEMBER TO ADD DEVID TO THE TOKEN
-    let mut rando = [0u8; 32];
-    rand::rng().fill_bytes(&mut rando);
-    let token = base64::engine::general_purpose::URL_SAFE.encode(&rando);
-    return token;
-}
-
-async fn revoketoken(pool: Data<PgPool>, token: &str) -> bool {
-    let _ = sqlx::query("DELETE FROM tokens WHERE token = $1")
-        .bind(token)
-        .execute(pool.get_ref())
-        .await
-        .map_err(|e| ErrorInternalServerError(e));
-    return true;    
-}
-
 
     /*sqlx::query("DELETE FROM tokens WHERE token = $1")
         .bind(token)
